+++ changelog.md
@@ -10,6 +10,7 @@ compromise uses semver, and pushes to npm frequently
 ##### 11.1.0
 - add `#Multiple` tag, and changes to how invalid numbers like 'sixty fifteen hundred' are understood
 - better em-dash/en-dash support
+- better conjugate implicit verbs inside contractions - "i'm", "we've"
 <!-- * include 'the #TitleCase' matches in .topics() -->
 
 ##### 11.0.0